+++ src/parser/tableref/subqueryref.cpp
@@ -39,7 +39,7 @@ unique_ptr<TableRef> SubqueryRef::Deserialize(Deserializer &source) {
 		return nullptr;
 	}
 	auto result = make_unique<SubqueryRef>(move(subquery));
-	int column_count = source.Read<int>();
+	size_t column_count = source.Read<int>();
 	for (size_t i = 0; i < column_count; i++) {
 		result->column_name_alias.push_back(source.Read<string>());
 	}
